Repurpose the create(_:) method to be store(_:)
Given the existence of a `create:` parameter in a Resource's
initializer, passing the instance method `create(_:)` to the initializer
parameter `store:` is very confusing, especially for people getting used
to the jargon of Vapor. This file is clearer now, but there's some
follow-up clarification to be done in the PostController's test cases.