Have `storeNewPost()` return the Post on success
Returning an optional Post-instance is closer to what you'd expect when
calling a method that converts a preset request to a Post and then
stores it. Even if the callsite is more interested in the Post's id.

This does make the map call to gather newly created ids very unreadable.